@@ -1,10 +1,16 @@
 module github.com/daishe/kubeconfig
 
-go 1.16
+go 1.20
 
 require (
-	github.com/chzyer/readline v0.0.0-20180603132655-2972be24d48e
-	github.com/manifoldco/promptui v0.8.0
-	github.com/spf13/cobra v1.1.3
-	golang.org/x/crypto v0.0.0-20210317152858-513c2a44f670
+	github.com/chzyer/readline v1.5.1
+	github.com/manifoldco/promptui v0.9.0
+	github.com/spf13/cobra v1.6.1
+	golang.org/x/crypto v0.6.0
+)
+
+require (
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+	golang.org/x/sys v0.5.0 // indirect
 )
